// web/src/apis/models/SubtitleSegment.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SubtitleSegment = {
    /**
     * The subtitle text
     */
    text: string;
    /**
     * Start time in seconds
     */
    start: number;
    /**
     * Duration in seconds
     */
    duration: number;
};