// web/src/apis/models/ContentMediaType.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum ContentMediaType {
    ARTICLE = 'article',
    VIDEO = 'video',
    TWITTER = 'twitter',
    PDF = 'pdf',
    WORD = 'word',
    EXCEL = 'excel',
    TEXT = 'text',
    PPT = 'ppt',
    AUDIO = 'audio',
    SPOTIFY_AUDIO = 'spotify_audio',
    AUDIO_INTERNAL = 'audio_internal',
    AUDIO_MICROPHONE = 'audio_microphone',
    IMAGE = 'image',
    TEXT_FILE = 'text_file',
}